Update express server example
